chore: Do queries in chart component
This gives us some more control over the query lifecycle (for showing
loading/error states) and cleans up the API a lot.
If we need the history outside of this component, we should be able to
just add another query hook and tanstack query should (?) de-dupe the
request for us.