fix: Stop tracking interval at the end of the current day
Include any stats from today in the tracking.

This greatly improves the experience if your stats have been registered
once before and you are returning. In this case we will now have 2 stat
points within the interval and be able to give estimates. Otherwise you
would have to come back another day, or tweak the interval to be in the
future.