chore: Point download links to github
Having an issue with chrome blocking the newest windows release, but it
seems to be tied to that binary, and not the fact that it is coming from
github. Point the download links to github again so I don't have to host
the files myself, and so that we still get the automated downloads
tracking that github does for us.