feat: Add an explicit "Search" option in the UserSearch component
There was an issue in the previous component with `autoHighlight`,
causing it to be impossible to search for a player's name if the name
matched another result. Since the other result was automatically
highlighted, it would navigate to that instead of searching for the new
name.

This fixes this issue by adding an explicit option at the bottom of the
list to search for the current value.