chore: Make eslint no-deprecated warning only
When working e.g. with dependabot it is nice to be able to upgrade
dependencies to a version where something is deprecated without having
to immediately act on it. That way we're sure we're on a version where
the recommended alternative is available on the main branch, and making
the change is easy :^)