chore: Don't duplicate workflow runs
Used to run both on push and pull request. This caused duplicate runs
when pushing to a branch with an open pull request.
Change this to run on all pull requests, and on pushes to main.

This won't run validation on pushes to other branches, but since I'm
mostly using a PR based workflow in this repo, this should be fine :^)