chore: Narrow `computeStat` type with overrides
We know that this function can only return `null` when querying
winstreaks. Encode this in the same way we have done it for `getStat`.

NOTE: Typescript does not seem to be able to detect when this overload
is being violated (e.g. by returning `null` for all stats). The same
issue seems to occur with `getStat`. We should be careful to uphold this
contract, as typescript will not help us.

Deemed worth it as the DX improves greatly, and these functions will
likely not be touched much.