chore: Drop "current" stats in progression calculation
This has been aliased to the end stats for a while, so we're making that
explicit now.
This was originally (potentially) two different stats instances, as the
tracking interval (used for progression rate calculations) did not
include the current time, mainly to make it more static.

This was later changed to actually include the current time to be able
to get more accurate progression rates, especially the second time
playing, or after long periods of inactivity.