fix: Handle infinite session quotient
We were making some comparisons between the session quotient and the
current and milestone quotients. These comparisons were incorrect when
there was no divisor increase in the session, as then the session
quotient is actually infinite, though we represent it as just the
session dividend.

Explicitly handle the case of infinite session quotient, and enable a
test case that caught this bug.